requirement.txt
torch
torchvision
streamlit
numpy
opencv-python-headless
Pillow
matplotlib
seaborn
scikit-learn
